Fix parse error on comment on import line
Minimal example:

    from x import * #
    a = 1

This commit adds a test and fix for this.

The root cause is that when `_get_token(skip_newline=False)` encounters
a comment, it will call itself without any argument, and so use the
default of `True` to skip next new line. The fix is to pass down the
correct argument.