run: Raise exception before printing error
If the command failed and was not allowed to, we should not print the
error message to stderr. The calling code will receive the same message
in an exception and should be free to decide whether it should be
printed somewhere.

The case when the command could fail is left unchanged. There is no easy
way to return the additional error message without changing the API of
the function. However this case can be ignored as any caller can mark
the command as non-failable and just ignore the exception.

Fixes: #25