log: Raise exception when log file is wrong
When the file can not be created for some reason, or is not writable, we
should not silently continue. There is no way for the user to tell if
the action they requested succeeded or not. This could easily lead to
logs being lost because the file was not created, yet no error reported.