Fix syntax error in base template
Bare auth/login doesn't make sense (auth variable divided by login?)
"auth/login" was clearly intended.

It seems that this couldn't possibly have worked for anyone, so that
probably means that all users of this code have deployed their own
base.html instead of using this default one (which makes sense).

The motivation for fixing it now is that I want to add autotests which
incidentally require rendering the template.