Use streaming response for task logs where possible
Based on comments and the use of _stream_file, it was intended
that these shouldn't load the entire log into memory.  That wasn't
accomplished; when HttpResponse is passed an iterable (which is
deprecated), it loads and joins every yielded chunk at once anyway.

StreamingHttpResponse is the correct response class to avoid this and
only load a chunk at a time.

This is a partial fix for kobo hub's handling of large logs.

This small part of the fix is being made first because it's cumbersome to
test the code otherwise: the Django test client calls close() on each
response before returning it to the caller.  On older Django versions,
when a closable iterable is passed to HttpResponse, and close() is called
before content is accessed, the content becomes unreadable (since reading
the content is delayed until first use).  Using StreamingHttpResponse as
intended avoids this problem and makes the code testable.