Avoid reading entire log file when loading from offset
Provides a new method for reading a specific chunk from a log file
without having to read the entire file into memory.

Currently, this reduces memory usage in the case where reading from
a certain offset is requested; log content prior to the offset will
no longer be loaded.  It also adds support for reading chunks in the
middle of log files (rather than reading to the end), but additional
commits are needed to take advantage of that.