Apply a max size to HTML log view
For large logs, displaying the entire log file is unsafe; it will
consume too much memory on the server and also make the browser
unresponsive.  A limit must be applied to make it safe.

The new behavior will show the end of the log only; users can click
through to "download" to access the entire log in raw format.

Introduces HTML_LOG_MAX_SIZE setting, which defaults to 2MB.