Let server control the LogWatcher refresh interval
LogWatcher is used to poll logs for updates.  Previously, the logic
was: if the task hasn't finished yet, then poll for a new update in 5
seconds.

It makes more sense for the server to control the poll interval.

Firstly, because it can be centrally configured by the new
LOG_WATCHER_INTERVAL setting.

Secondly, because there can be cases other than "task isn't finished
yet" which still require another request to the server for log content;
in particular, an upcoming commit will adjust the API to read logs in
chunks, so additional requests may be needed to read more chunks even if
the task already finished.

This change is written so that old JS on new server will still work.
That's necessary since some applications have deployed this JS by
copying it into their own sources.

The watch-log command also needs to inspect the next_poll attribute to
determine whether it should poll again, though it does not respect the
value provided by the server, instead using the given command-line
option.