Make JSON API for logs return chunks
This would previously return entire log content from the requested
offset to the end of the file.  That could be too large to fit in
memory, so apply a limit (controlled by JSON_LOG_MAX_SIZE setting).

When the available log content exceeds the log size, the client is
instructed via next_poll to request more content soon.

Note that there's still no client-side limit on the amount of data
fetched, so viewing a large log may still result in a lagging
browser.