Improve python 2.6 compatibility, portability of tests
A few tests used unittest features introduced in Python >= 2.7.
Use unittest2 module from pypi to get these features also in Python 2.6.

Two tests require rpm and koji modules to be installed, but the upstream
projects for these do not package them as python eggs or ship them to
pypi, making them impractical to install into a virtualenv.  That means
there's a high chance the modules won't be available; in that case, skip
those tests.

Tweak the test runner script so there's a visual indicator of skipped
tests.

The goal is to make "tox" a likely-to-succeed method of running
autotests for developers, with a low rate of false positives.