Revise log trimming to show end of log, not beginning
When viewing a log larger than HTML_LOG_MAX_SIZE, the previous behavior
was to render a "log is trimmed" message and the first chunk of the log.

It's more useful to show the end of the log instead.  This is
particularly important when the task is still in progress, and is more
consistent with how other systems will stream large task logs (e.g.
Jenkins).

Note that this also removes a feature: "offset" parameter is no longer
accepted when using an HTML log view.  Since there's no clear use-case
for it and none of the views ever generated any links with this parameter,
I think this is OK.