Use buffered readers when accessing log files
In the new "tail" method of showing the end of log files, we need to do
line-based reading of the entire log, which can be slow - particularly
in the gzip case.  Buffered reading makes it significantly faster.

This change reduces runtime of test_view_log.py by about 50%.

Unfortunately python 2.6 is unable to make full use of buffered reading,
so will have worse performance on gzipped logs.