six is a runtime requirement
Prior to this change, "pip install kobo" would not install "six". This
is a problem because we depend on six in a number of locations at
runtime. For example "import kobo.rpmlib" would fail because nothing
pulled in six.

The Tox tests were passing because Tox installed six as part of
the test operation (tox.ini).

Move the six package from tox.ini to the setuptools packaging so all
kobo users will install six.