Limit dashes in log files
When the command is echoed in the log file, it has an underline of
dashes, that matches the length of the command. This is nice as a
delimiter, but for very long commands it can be a little annoying. Let's
limit the maximum length to 79, which should fit on a reasonably sized
terminal without wrapping.