tests: ditch custom test runner in favor of pytest
Get rid of the custom run_tests.py wrapper and use py.test to run
tests instead.  The primary motivation is to allow us to benefit
from the available pytest features and plugins, such as the ability
to make an XML test report or to use pytest-cov for measuring
coverage.

Includes related changes:

- fold the tests/README.md info into the main README and make it
  up-to-date

- move test deps from tox.ini to test-requirements.txt so they
  can be installed when not using tox

- make tests work regardless of working directory when test is
  started

- remove ad-hoc overriding of PYTHONPATH

- remove some unnecessary code for dealing with django settings

- port a couple of uses of django test runner to pytest
  {setup,teardown}_module hooks