Stop decoding sigkey headers
The values are actually binary data that has to be processed in a
complicated way. Decoding it as UTF-8 is wrong.

Instead of more magic to determine what to decode or not, this patch
adds an explicit option.

The unittest is using headers from magic-8.2.151-1.fc30.x86_64.rpm
Fedora RPM signed by actual release keys.