Improve compatibility for newer versions of Django
* Staticfiles app has been added to Django for providing
static page elements.

* Old, deprecated krbV library has been replaced by gssapi.
The new library should work with both Python 2 and 3.
The authentication process attempts to be as close to the
original one as possible.

* SubfieldBase inheritence has been removed, which means that six
metaclass workaround could be removed. Inhereted 'from_db_value'
method had to be reimplemented to not cause any functionality loss.

* Classes representing HTML elements now return their representation
as __str__ instead of __unicode__ to be compatible with Python 3.
Python 2 compatibility is ensured by @python_2_unicode_compatible
decorator.

* Django's URL paths are not written as strings anymore, as this
isn't supported by the newer Django versions.

* SortedDict object has been replaced by OrderedDict, because it
was deprecated in newer versions of Django.

* Parameters of the RemoteUserBackend's authenticate method have
changed in Django 1.11

* Python 3 doesn't support unbuffered text output, which is why
TextIOWrapper workaround was used in Python 3