Fix typo and remove a faulty test
The believed reason that the test 'test_without_default' has started
to fail is the change of behavior of an assignment caused by the
deprecation of the class SubfieldBase. With the removal of
'SubfieldBase', the method 'to_python' is
no longer being called upon construction [1], which results in
the expected Exception no longer being raised.

[1] https://code.djangoproject.com/ticket/26807