Add gssapi as a kerberos authentication method
Since krbV is not compatible with Python 3, a new login
method has been added which uses the gssapi library. The
new method is not a replacement for the old one, and will
not be used by default in order to ensure backwards compatibility.
The new library should be compatible with Python 2 and 3.