Fix Django migrations
Migration process was broken in Python 3 due to a Django bug [1].
Esentially, a migration file was missing which would alter all
binary data to strings. This happens when a migration is generated
in Python 2 and attempted to be run in Python 3. A better solution
is to alter the original migration to use string data from the get-go.
It was tested on Python 2.7 with Django 1.8, and Python 3 with
Django 1.11. This operation is considered safe in the bug discussion.

[1]: https://code.djangoproject.com/ticket/23226