Fix Django testcases
Many tests were failing because Django database transaction doesn't
get flushed after a test failure (even marked with xfail). This prevents
the setup of the next test to be conducted properly because to-be-created
data are already present in the database. The root cause is unknown
(perhaps a bug that was fixed in later versions of Django?). It was
fixed by changing the inherited testing class from 'django.test.TestCase'
to 'django.test.TransactionTestCase'. They should be functionally identical,
they only differ in how the Database cleanup gets performed. Additionally,
a '_fixture_teardown()' method was added to the test class's setUp method
to ensure that the database has been flushed before the next test.

Testcase 'test_get_worker_id' was broken by these changes, as the ID
of the created worker is no longer 1. This is most likely caused
by the difference in how the objects get removed from the database.
Testcase was instead changed to check if any viable ID is present.