Add kerberos authentication via GSSAPI
This new authentication method is meant to replace krbV, which is
not compatible with Python 3. After the kerberos authentication is
complete, Django authentication is performed, just like in 'login_krbV'.
To be gssapi compliant, the authentication is performed in a loop
(despite the fact that kerberos should only require one step).

Also, XMLRPC view was broken in the newer Django versions (1.11),
so a fix has been added as well.