Change 'render_to_response' to 'render'
'render_to_response' is deprecated in favor of 'render'[1].
'render' should be available in the older verions of Django
as well, so multiple code paths are not necessary. Also,
'context_instance' parameter is deprecated and was replaced
by request as described in the docs[2].

'render_to_response' is still being used in the 'SearchView'
class methods. It's not a 'django.shortcuts' method, but it's
a method inherited from ExtraListView's ancestor 'TemplateResponseMixin',
and is still present in Django 3.0.

[1] https://docs.djangoproject.com/en/3.0/releases/2.0/#id1
[2] https://docs.djangoproject.com/en/2.2/releases/1.8/#dictionary-and-context-instance-arguments-of-rendering-functions