Ensure that 'BCryptPasswordHasher' is available in new Django versions
'BCryptPasswordHasher' was removed from default password hashers
in Django 2.1[1]. As kobo might work with historical data,
PASSWORD_HASHERS which included this hasher along with other
default hashers[2][3] were explicitly defined in the settings.
The two definitions based on the Django version are necessary
because some of the newer default hashers are not available
on old Django versions.

[1] https://docs.djangoproject.com/en/3.0/releases/2.1/#removed-bcryptpasswordhasher-from-the-default-password-hashers-setting
[2] https://docs.djangoproject.com/en/1.8/ref/settings/#password-hashers
[3] https://docs.djangoproject.com/en/3.0/ref/settings/#password-hashers