Fix backwards-incompatible MIDDLEWARE check
The logic here was: if Django 1.10 or greater is used, then
check the MIDDLEWARE variable.  Else check MIDDLEWARE_CLASSES.

This isn't correct, as Django 1.10 added support for MIDDLEWARE
but did not remove support for MIDDLEWARE_CLASSES.  This was a
backwards-incompatible change, as an application using
MIDDLEWARE_CLASSES could work fine on kobo 0.14.0 but would start
crashing on kobo 0.15.0, for no good reason.

Fix it so it's backwards-compatible. There are Django versions
supporting both MIDDLEWARE or MIDDLEWARE_CLASSES, so the only way
we can know which one to check is to probe which one has been
defined.