HubProxy: remove broken auto_logout feature
This class had an implementation of __del__ to automate a call
to logout when instances are finalized.  This is not a good idea
for a few reasons:

- there's no guarantee that __del__ is ever called, so if the logout
  here was important, this wouldn't be an appropriate way to ensure
  it happens

- __del__ is called at unpredictable times, as GC can be triggered
  anywhere, so doing non-trivial things such as HTTP requests is
  prone to bugs. Example: if GC starts while locks are held, the
  call to logout() itself may attempt to grab a lock which is
  already held, thus causing deadlock. This is known to happen
  with django Signal.lock when autotests are using the Django
  test client.

- it appears unimportant. The impact of logout() is to remove a
  session from the server, but the server obviously has to deal with
  expiring sessions with no explicit logout anyway.

If this feature is needed, it could be implemented by making HubProxy
a context manager which logs out during __exit__, but it is thought
to be not useful.