Fix kobo.django.xmlrpc.views
django.template.loader.get_template returns
django.tepmlate.backends.django.Template. It differs from
django.template.Template in api of its render() function (not only).
This leads to error when the render function of django.template.Template
is called with wrong arguments.

The solution is to call the render function of each type of template
differently.