Process packages with header only signature
An RPM consists of four parts: lead, sighdr, hdr and payload.

When `rpm --addsign` runs, it computes two signatures: a `sigt_v3`,
which is a signature over `hdr || payload` (`||` is bit-wise appending),
and a `sigt_v4`, which is a signature over only `hdr`. It's fine to sign
only the header, because it contains a checksum of payload.

Header-only signing means skipping `sigt_v3` and only building
`sigt_v4`. This avoids having to send the full payload over to the
signing server.

RPM has been using both methods for a long time, and since 4.16, it only
does `sigt_v4` (header-only signing) by default, and `sigt_v3` has to be
requested specifically.