Change the ordering of hub tasks in UI [RHELDST-8253]
When viewing the finished tasks list, the hub tasks were sorted by the
finished date. This led to the behaviour that the tasks that did not
finish (because they were cancelled) always sorted above all other
tasks and cluttered the finished task listing. We thus change the
ordering to use "dt_created" instead.