Fix LoggingThread randomly crashing due to mixing bytes & text
As kobo replaces sys.stdout, sys.stderr with an object pointing at
LoggingThread, this means LoggingThread's queue will receive whatever
type of object is passed into sys.stdout.write().

Problem: on python 2, this is allowed to include both bytes and text,
since that's how sys.stdout works there. However, if the queue
contains both at the same time, then this logic:

    while True:
        try:
            self._send_data += self._queue.get_nowait()
        except queue.Empty:
            break

... means implicit de/encoding will occur, since we might try to
concatenate bytes and unicode objects. If the implicit de/encoding is
using ASCII and if the queue contains a unicode object having
characters not able to be represented in ascii, a crash would occur.

For example, while neither of these crash:

    # case 1
    >>> x = ""
    >>> x += u'hi'

    # case 2
    >>> x = ""
    >>> x += b"\xe2"

This will crash:

    >>> x = ""
    >>> x += u'hi'
    >>> x += b"\xe2"
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    UnicodeDecodeError: 'ascii' codec can't decode byte 0xe2 in position 0: ordinal not in range(128)

This in turn means that e.g. the following may crash LoggingThread
when the stdout redirection is in place:

    >>> sys.stdout.write(u'hi')
    >>> sys.stdout.write(b"\xe2")

Why only "may crash": because it depends on how fast the thread is
reading items from the queue. If it's fast enough so that the above
loop tends to read only a single item before breaking, it would not
crash. The crash only occurs if the thread happened to read bytes
and text together during that loop.

Coupled with the fact that errors in the LoggingThread don't get written
anywhere obvious, this could give the impression that logging would
randomly and mysteriously stop an hour or more into a long-running task.

I'm not certain if this issue also affected python 3, as one is always
supposed to be writing text to sys.stdout there by default. However,
there are some ways to configure sys.stdout to receive bytes also in
python 3, so it seems prudent to keep this "convert to bytes" logic
permanently and not just as a py2-specific workaround.