Create URLs for OIDC-based authentication
OIDC-based authentication works similiarly to Kerberos, where a URL has
to be protected by a web server's authentication module, which upon
successful authentication sets the REMOTE_USER environment variable.
This variable is then used by Django to log the user in, as described
in[1]. The solution was verified to work with Apache and mod_auth_openidc[2].

The flow can be roughly described as follows:
- User accesses the <hostname>/auth/oidclogin/ endpoint protected by
  mod_auth_openidc
- Denied access initiates the authentication
- OIDC uses vanity URL also protected by mod_auth_openidc to complete
  the authentication
- Successful auth ends at URL <hostname>/auth/oidclogin/, where it sets
  the REMOTE_USER variable
- Django's middleware and backend logs the user in
- <hostname>/auth/oidclogin/ can now be accessed, and automatically
  redirects the user to the homepage.

A small limitation was introduced: URL parameters in the protected URL
cannot be used with OIDC auth, as the auth server itself attempts to
interpret them, causing 500 errors. Login normally uses the "next"
parameter, denoting where to redirect the user after a sucessful login
(URL where the login was clicked). As this argument cannot be passed for
OIDC auth, login will always redirect the user to the homepage.

Another small change from Kerberos auth is that a 302 (temporary)
redirect has to be used on the login URL, because 301 would be cached by
the browser and used on subsequent login attempts. This would
effectively break the login, as protected endpoint would not be
used directly.

[1] https://docs.djangoproject.com/en/4.0/howto/auth-remote-user/
[2] https://github.com/zmartzone/mod_auth_openidc