Implement OIDC-based authentication for clients
The client authenticates with the authentication server with Kerberos,
and once authenticated receives a JWT from the server. Then, client
sends the JWT to the hub, which uses it to authenticate the user.
The hub tries to decode the JWT with a public key received from the
auth server, and verify that the JWT is not expired and that it was
generated for the correct app. If everything checks out, the client
is logged in and is able to perform restricted actions via the HubProxy.

Hub uses cache so that it doesn't have to query the auth server for the
public key in every request. The cache expires after 1 day.

A limitation of this solution is that the CLI client has to query the auth
server for each action it performs, which may introduce significant
overhead to the server. As far as I know, large amounts of commands done
via a CLI in a short time timespan is not a typical use-case, so the
introduction of OIDC auth shouldn't cause any issues for the auth
server.