Implement OIDC-based client authentication
Authentication is performed by querying the hub endpoint protected by
mod_auth_openidc. OIDC authenticates the user with their Kerberos
ticket, and redirects back to hub. Django then recognizes REMOTE_USER
set by mod_auth_openidc and logs the user in. The user's login is
captured in the session's cookiejar. After login is complete, the cookie
is added to the XMLRPC transport object, which uses it to authenticate
when making requests to hub via XMLRPC.