Add option to use a custom task manager class
The class is passed a parameter to the main function. If a parameter is
passed, this class is used as a task manager instead of the default one.
Base class declaring which methods must the task manager implement was
added as well. These methods are called by the main loop and govern the
workflow of the task manager.