shortcuts.run: Use backslashreplace
If `errors` is `None`, the default data-to-text decoding error
handler is `'strict'`. This makes the user code to fail if the
stream contains a byte sequence that cannot be converted to text
under selected encoding. By setting `errors='backslashreplace'`
such a sequence of bytes will be replaced by escape sequences
instead of raising the exception.

Context: As a `kobo` consumers (covscan), we agree that it will be
better for us to have this change in `kobo` instead of adding
`errors='backslashreplace'` to many places in our code or implementing
a wrapper and backporting the patches to production branches.