Decode XMLRPC chunk regardless of its type
Currently, the decode function is decided based on the
Python version (decodestring for PY2, decodebytes for PY3). This
generally works, since chunk has bytes type on  Python 3
and str type on Python 2. However, if the XMLRPC client runs on Python
2 and the server runs on Python 3, the chunk will have the str type,
causing the decodebytes function to fail. Replace it instead with the
b64decode function, which is available since Python 2.4 and doesn't care
about the type of its input.