Addressing bandit findings and add bandit [RHELDST-12089]
We're adding in the Bandit scan in order to satisfy the ESS guidline
SEC-APP-REQ-2.

There are two points where nosec have been used that should be noted.
First is the subprocess.Popen in shortcuts.run. While adding Bandit
scan to rcm-pub we discussed this function and concluded it should be
left unchanged. This is because it's widely used and changing it would
outright break many tools which use this library. The other nosec comment is
on HTTPSConnection in http.py. This is because the issue affects old versions
of python 2.7 and python 3.4. I think it would be unlikely for this code to
be used with a vulnerable interpreter, and I've struggled to find other options
for python2. If this should beaddressed and there are actually viable options
I'll happily revist this.