Fix calculating the number of task's subtasks in new Django versions
Task's subtasks() method find all tasks whose parent is the current
task. The subtask_count field is set before the current task is saved to
the DB, meaning that subtasks() method matches all tasks without a
parent task, which is not the intended behavior. Fix setting
subtask_count by first saving the task to the DB so that subtasks()
correctly matches the parent task.

This issue is not present when an old version of Django is used.