Add Client Credentials Flow OIDC authentication
The authentication is primarily for automated use-cases that use service
accounts to log in. Client receives the access token from SSO using its
ID and secret. It then sends the access token to the hub, which
validates it and logs the user in. The newly added tokenidclogin URL is
expected to be protected by mod_oauth2 (or analogous for the used web
server), which handles the validation and the passing of REMOTE_USER to
Django. XML-RPC client uses the sessionid cookie (which ensures that the
user is logged in) provided by Django.