Don't raise an error when a character cannot be encoded/decoded
An issue was encountered where task logs failed to be decoded, causing
an error to be raised. When an unknown character is encountered, replace
it with ? instead of raising an error.

Also, Django was removed from test-requirements.txt, since it's supplied
by tox, and its duplication causes the installation to fail on lower
Python versions.