Make kobo compatible with Django 4.2
The only change that was required is the adding of updated fields to
update_fields parameter in the custom save methods[1]. Django's
get_or_create method now fills the update_fields parameter when calling
model's save method. This parameter ensures that only the specified
model fields will be updated. If a custom save method modifies
additional fields, this would result in them not actually being saved to
the DB. So, if save was called with the update_fields parameter, extend
it with all the modified fields.

[1] https://docs.djangoproject.com/en/4.2/releases/4.2/#setting-update-fields-in-model-save-may-now-be-required