hub: support creation of subtasks with an inherited worker from parent task
OpenScanHub expects the parent task and its subtasks to be executed on the
same worker.  However, there is a race with the `assign_task` XML-RPC call
because it does not guarantee that the subtask won't be assigned to another
worker by the hub in the meantime.

This commit adds a support for creation of subtasks pre-assigned to the
same worker executing the parent task, thus, invalidating the necessary
condition for the race described above.

Related: https://github.com/openscanhub/openscanhub/issues/156