Make LoggingThread recover on all errors [RHELDST-332]
Previously, LoggingThread would recover from any XML-RPC fault, but
would stop when any other type of exception was encountered.
That is a problem, as it means the worker will permanently give
up sending messages to the hub when all kinds of temporary
issues occur (e.g. a temporary network disruption between worker
and hub). The task underneath may continue running for hours,
with all log messages being discarded.

Given the nature of this thread, it makes more sense to attempt
recovering from all kinds of errors, as we should try hard not
to lose log messages from a task.

Fixes https://github.com/release-engineering/kobo/issues/60

(This commit is a reimplementation of
https://github.com/release-engineering/kobo/pull/106)