worker: finish all opened tasks and reexec on SIGHUP
This feature implements safe reloading of updated configuration or updated
source code, e.g. with `systemctl reload`.

The reexec won't be performed when SIGHUP is followed by SIGTERM, SIGINT
or explicit shutdown by the `ShutdownWorker` task.

The unit tests must use `mock.ANY` instead of the actual signal handler
because it is impossible to reference a nested closure outside of its own
scope.