pkgset.FileCache: implement __contains__ for efficiency
See https://pagure.io/pungi/pull-request/1796 . We figured out
there that `foo in somefilecache` works, but is much less
performant than `foo in somefilecache.file_cache`, because the
latter hits the efficient `__contains__` method that Python
dicts have, but the former winds up iterating over the dict's
keys via `FileCache.__iter__`. In an extreme case, this was making
something take 20 minutes instead of 3 seconds.

Users can avoid this by just hitting the `file_cache` directly
as in the PR, but to fix this for any other users without them
having to change, let's implement `__contains__` in `FileCache`
by just passing it through to the dict.

Signed-off-by: Adam Williamson <awilliam@redhat.com>