Add transport args as a kwarg for HubProxy [RHELDST-27818]
We would like to be able to set the retry parameters for xml-rpc calls
in pub. The retry_request decorator accepts kwargs for setting the retry
parameters, but the HubProxy class doesn't make use of these.

This change allows users to provide transport class args as a kwarg, so
we can modify the retry behaviour.