hub: use .save() with enable-worker so that ready state is refreshed
When enable-worker and disable-worker commands are used, they update
the enabled field in Worker objects directly, bypassing the .save()
method, which is the only place where the ready field is recalculated.

This change uses .save() to make the enabled state change, so that
ready state is recalculated. This fixes an issue where if a worker is
disabled, then later re-enabled, it can get stuck in a non-ready
state.